Default Game: improved save slot lookup
With original script an existing slot could be overwritten with a new save name if some of the early save files are missing.
For example, if the game saves folder has following saved games:

agssave.003 -- named "Save 3"
agssave.004 -- named "Save 4"

User types in "New Save", game cannot find such slot and assigns slot number to the number of saves + 1, that is: 3. And the game is written to the existing file named agssave.003, overwriting game which we did not intend to overwrite.