Remove keyboard shortcut for opening inventory
The operation of the inventory GUI currently doesn't mix well with being
opened with the keyboard when there is an active inventory item. Rather
than add more complication and tracking to the inventory GUI handling
this just removes the keyboard shortcut for the time being. Ideally the
inventory GUI behaviour should be improved or be made a more modular
component.