BASS: Fix inventory right-click with active item
Changes behaviour when right-clicking on an inventory item whilst the
player has an active inventory item. The active inventory item is now
deselected instead of running the look interaction on the item which
was right-clicked.

This matches the operation of the older version of the BASS template
before it was rewritten.